# Conflicts:
#	packages/server/forum/controllers/topicController.ts
#	packages/server/forum/init.ts

+++ packages/server/forum/controllers/topicController.ts
@@ -6,22 +6,25 @@ import MessageModel from '../models/messageModel';
 
 const getTopics = (_: Request, res: Response) => {
   TopicModel.findAll({
-    order: [[Sequelize.col('TopicModel.createdAt'), 'DESC']],
+    order: [[Sequelize.col('createdAt'), 'DESC']],
   })
-    .then((topic) => {
-      res.status(200).json(topic);
+    .then((topics) => {
+      res.status(200).json(topics);
     })
-    .catch((error) => {
-      res.status(400).json({ message: error });
+    .catch(() => {
+      res.status(400).json({ message: 'Bad request' });
     });
 };
 
 const postTopic = (req: Request, res: Response) => {
+  const { user } = res.locals;
+  const { id, login, avatar } = user;
+
   TopicModel.create({
     ...req.body,
-    owner_avatar: 'TEST AVATAR',
-    owner_id: 'TEST ID',
-    owner_login: 'TEST LOGIN',
+    owner_avatar: avatar,
+    owner_id: id,
+    owner_login: login,
   })
     .then((topic) => {
       res.status(200).json(topic);
@@ -52,36 +55,56 @@ const getTopicWithMessages = (req: Request, res: Response) => {
 };
 
 const updateTopic = (req: Request, res: Response) => {
-  const { id } = req.params;
-  TopicModel.update(
-    {
-      ...req.body,
-      owner_avatar: 'SUPERTEST AVATAR',
-      owner_id: 'SUPERTEST ID',
-      owner_login: 'SUPERTEST LOGIN',
-    },
-    { where: { id }, returning: true },
-  )
-    .then((topic) => {
-      res.status(200).json(topic[1][0]);
-    })
-    .catch(() => {
-      res.status(400).json({ message: 'Bad request' });
-    });
+  const { user } = res.locals;
+  const { id, login, avatar } = user;
+
+  TopicModel.findOne({ where: { id } }).then((topic) => {
+    if (topic && String(user.id) === topic.owner_id) {
+      TopicModel.update(
+        {
+          ...req.body,
+          owner_avatar: avatar,
+          owner_id: id,
+          owner_login: login,
+        },
+        { where: { id }, returning: true },
+      )
+        .then((topic) => {
+          res.status(200).json(topic[1][0]);
+        })
+        .catch(() => {
+          res.status(400).json({ message: 'Bad request' });
+        });
+    } else {
+      res.status(403).json({ message: 'Wrong owner' });
+    }
+  });
 };
 
 const deleteTopic = (req: Request, res: Response) => {
   const { id } = req.params;
-  TopicModel.destroy({
-    where: { id },
-  })
-    .then((deletedRecord) => {
-      deletedRecord === 1
-        ? res.status(200).json({ message: 'OK' })
-        : res.status(404).json({ message: 'Not found' });
+  const { user } = res.locals;
+
+  TopicModel.findOne({ where: { id } })
+    .then((topic) => {
+      if (topic && topic.owner_id === String(user.id)) {
+        TopicModel.destroy({
+          where: { id },
+        })
+          .then((deletedRecord) => {
+            deletedRecord === 1
+              ? res.status(200).json({ message: 'OK' })
+              : res.status(404).json({ message: 'Not found' });
+          })
+          .catch(() => {
+            res.status(400).json({ message: 'Bad request' });
+          });
+      } else {
+        res.status(403).json({ message: 'Wrong owner' });
+      }
     })
     .catch(() => {
-      res.status(400).json({ message: 'Bad request' });
+      res.status(400).json({ message: 'not found' });
     });
 };
